+++ internal/pgputil/suite_test.go
@@ -1,11 +1,11 @@
-package openpgputil_test
+package pgputil_test
 
 import (
 	"testing"
 
 	. "gopkg.in/check.v1"
 
-	"github.com/canonical/chisel/internal/openpgputil"
+	"github.com/canonical/chisel/internal/pgputil"
 )
 
 func Test(t *testing.T) { TestingT(t) }
@@ -15,11 +15,11 @@ type S struct{}
 var _ = Suite(&S{})
 
 func (s *S) SetUpTest(c *C) {
-	openpgputil.SetDebug(true)
-	openpgputil.SetLogger(c)
+	pgputil.SetDebug(true)
+	pgputil.SetLogger(c)
 }
 
 func (s *S) TearDownTest(c *C) {
-	openpgputil.SetDebug(false)
-	openpgputil.SetLogger(nil)
+	pgputil.SetDebug(false)
+	pgputil.SetLogger(nil)
 }
