+++ internal/setup/setup.go
@@ -13,7 +13,7 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/canonical/chisel/internal/deb"
-	"github.com/canonical/chisel/internal/openpgputil"
+	"github.com/canonical/chisel/internal/pgputil"
 	"github.com/canonical/chisel/internal/strdist"
 )
 
@@ -427,7 +427,7 @@ func parseRelease(baseDir, filePath string, data []byte) (*Release, error) {
 	// Decode the public keys and match against provided IDs.
 	pubKeys := make(map[string]*packet.PublicKey, len(yamlVar.PubKeys))
 	for keyName, yamlPubKey := range yamlVar.PubKeys {
-		key, err := openpgputil.DecodePubKey([]byte(yamlPubKey.Armor))
+		key, err := pgputil.DecodePubKey([]byte(yamlPubKey.Armor))
 		if err != nil {
 			return nil, fmt.Errorf("%s: cannot decode public key %q: %w", fileName, keyName, err)
 		}
